requirements.txt
selenium
pytest
pytest-xdist
pytest-html
allure-pytest
openpyxl
pyyaml
